Extract CodyWindowAdapter
