+++ src/main/kotlin/com/sourcegraph/cody/config/CodyAuthenticationManager.kt
@@ -18,9 +18,7 @@ import com.intellij.openapi.util.Disposer
 import com.intellij.openapi.wm.WindowManager
 import com.intellij.util.AuthData
 import com.intellij.util.concurrency.annotations.RequiresEdt
-import com.sourcegraph.cody.agent.CodyAgent
 import com.sourcegraph.cody.agent.CodyAgentService
-import com.sourcegraph.cody.agent.protocol_generated.Window_DidChangeFocusParams
 import com.sourcegraph.cody.api.SourcegraphApiRequestExecutor
 import com.sourcegraph.cody.api.SourcegraphApiRequests
 import com.sourcegraph.cody.config.notification.AccountSettingChangeActionNotifier
@@ -28,8 +26,6 @@ import com.sourcegraph.cody.config.notification.AccountSettingChangeContext
 import com.sourcegraph.cody.config.notification.AccountSettingChangeContext.Companion.UNAUTHORIZED_ERROR_MESSAGE
 import com.sourcegraph.config.ConfigUtil
 import java.awt.Component
-import java.awt.event.WindowAdapter
-import java.awt.event.WindowEvent
 import java.util.concurrent.CompletableFuture
 import java.util.concurrent.Executors
 import java.util.concurrent.TimeUnit
@@ -111,23 +107,7 @@ class CodyAuthenticationManager :
 
   fun addAuthChangeListener(project: Project) {
     val frame = WindowManager.getInstance().getFrame(project)
-    val listener =
-        object : WindowAdapter() {
-          override fun windowActivated(e: WindowEvent?) {
-            super.windowActivated(e)
-            ApplicationManager.getApplication().executeOnPooledThread { getAuthenticationState() }
-            CodyAgentService.withAgent(project) { agent: CodyAgent ->
-              agent.server.window_didChangeFocus(Window_DidChangeFocusParams(true))
-            }
-          }
-
-          override fun windowDeactivated(e: WindowEvent?) {
-            super.windowDeactivated(e)
-            CodyAgentService.withAgent(project) { agent: CodyAgent ->
-              agent.server.window_didChangeFocus(Window_DidChangeFocusParams(false))
-            }
-          }
-        }
+    val listener = CodyWindowAdapter(project)
     frame?.addWindowListener(listener)
     Disposer.register(this) { frame?.removeWindowListener(listener) }
   }
@@ -135,7 +115,7 @@ class CodyAuthenticationManager :
   @CalledInAny fun getAccounts(): Set<CodyAccount> = accountManager.accounts
 
   @CalledInAny
-  private fun getAuthenticationState(): AuthenticationState {
+  fun getAuthenticationState(): AuthenticationState {
     val previousIsTokenInvalid = isTokenInvalid?.getNow(null)
     val previousTier = tier?.getNow(null)
     val isTokenInvalidFuture = CompletableFuture<Boolean>()
